requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
pymupdf
python-docx
python-dotenv
requests
langchain